feat!: move redis memory to one file
