+++ Cargo.toml
@@ -8,7 +8,6 @@ bench = []
 
 [dependencies]
 aes = "0.8.4"
-colored = "2.1.0"
 rand = "0.8.5"
 rand_chacha = "0.3.1"
 rand_core = "0.6.4"
